requirements.txt
accelerate==0.29.3
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.4
Django==5.0.4
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
huggingface-hub==0.22.2
idna==3.7
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-text-splitters==0.0.1
langsmith==0.1.49
MarkupSafe==2.1.5
marshmallow==3.21.1
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
orjson==3.10.1
packaging==23.2
psutil==5.9.8
pydantic==2.7.0
pydantic_core==2.18.1
python-decouple==3.8
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
SQLAlchemy==2.0.29
sqlparse==0.5.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.2
transformers==4.40.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
